[#21] refactor : 이름 수정, app/globals.css 제거, 태그 개선
