+++ app/create/(route)/petgroup/page.tsx
@@ -1,10 +1,9 @@
 "use client";
 
 import { NextPage } from "next";
-import "@/app/globals.css";
 import * as styles from "./page.css";
 
 const CreatePetGroupPage: NextPage = () => {
-  return <section className={styles.container}>펫 프로필 생성 페이지</section>;
+  return <main className={styles.container}>펫 프로필 생성 페이지</main>;
 };
 export default CreatePetGroupPage;
