+++ app/create/(route)/userprofile/page.tsx
@@ -1,10 +1,9 @@
 "use client";
 
 import { NextPage } from "next";
-import "@/app/globals.css";
 import * as styles from "./page.css";
 
 const CreateUserProfilePage: NextPage = () => {
-  return <section className={styles.container}>유저 프로필 생성</section>;
+  return <main className={styles.container}>유저 프로필 생성</main>;
 };
 export default CreateUserProfilePage;
