+++ components/ParticipatePetGroupModal/ParticipatePetGroupModal.tsx
@@ -1,14 +1,13 @@
 import { NextPage } from "next";
-import "@/app/globals.css";
-import * as styles from "./PariticipatePetCroupModal.css";
+import * as styles from "./ParticipatePetGroupModal.css";
 
 const ParticipatePetGroupModal = () => {
   return (
-    <main className={styles.backDropContainer}>
+    <section className={styles.backDropContainer}>
       <section className={styles.modalContainer}>
         <p>전달 받으신 반려동물 등록 코드를 입력해주세요.</p>
       </section>
-    </main>
+    </section>
   );
 };
